move NPLINKER_APP_DATA_DIR to defaults to avoid partial initialisation
